requirements.txt
psycopg2-binary
pandas
openpyxl
python-dotenv
xlrd
fastapi
uvicorn
jinja2